购物车的变量处理

+++ trade-center/trade-center-servie/src/main/java/com/ai/bdex/dataexchange/tradecenter/service/interfaces/order/IOrdMainInfoSV.java
@@ -9,7 +9,7 @@ import com.ai.bdex.dataexchange.tradecenter.dubbo.dto.order.OrdMainInfoReqDTO;
 import com.ai.bdex.dataexchange.tradecenter.dubbo.dto.order.OrdMainInfoRespDTO; 
 public interface IOrdMainInfoSV { 
 	 //创建订单
-	 public int creatOrderByweb(OrdMainInfoReqDTO ordMainInfoRespDTO)throws Exception;
+	 public long creatOrderByweb(OrdMainInfoReqDTO ordMainInfoRespDTO)throws Exception;
 	 //根据ID查询订单的详细信息
 	 public OrdMainInfo queryOrderById(OrdMainInfo ordInfo) throws Exception;
 	 //我的订单 （可以根据状态查询是否已经申请开票了）
